Search moved under Documents index action. Old index action moved to Details
